requirements.txt
asgiref==3.7.2
asyncpg==0.28.0
backports.zoneinfo==0.2.1
bleach==6.0.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
crispy-bootstrap4==2022.1
cryptography==41.0.2
defusedxml==0.7.1
Django==4.1
django-allauth==0.54.0
django-cors-headers==4.2.0
django-crispy-forms==2.0
flake8==6.1.0
idna==3.4
isort==5.12.0
mccabe==0.7.0
mysqlclient==2.2.0
oauthlib==3.2.2
Pillow==10.0.0
psycopg2==2.9.6
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.7.0
python-dotenv==1.0.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==1.26.6
webencodings==0.5.1
whitenoise==6.5.0
